requirements.txt
streamlit
streamlit-mic-recorder
google-generativeai
python-dotenv